# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "RetailConnect",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/campaigns/{campaignId}/touchpoints/{id}":  {
                                                                       "delete":  {
                                                                                      "tags":  [
                                                                                                   "CampaignTouchpoint"
                                                                                               ],
                                                                                      "parameters":  [
                                                                                                         {
                                                                                                             "name":  "id",
                                                                                                             "in":  "path",
                                                                                                             "required":  true,
                                                                                                             "schema":  {
                                                                                                                            "type":  "integer",
                                                                                                                            "format":  "int32"
                                                                                                                        }
                                                                                                         },
                                                                                                         {
                                                                                                             "name":  "campaignId",
                                                                                                             "in":  "path",
                                                                                                             "required":  true,
                                                                                                             "schema":  {
                                                                                                                            "type":  "string"
                                                                                                                        }
                                                                                                         }
                                                                                                     ],
                                                                                      "responses":  {
                                                                                                        "200":  {
                                                                                                                    "description":  "Success"
                                                                                                                }
                                                                                                    }
                                                                                  }
                                                                   },
...
                  "/api/touchpoints":  {
                                           "post":  {
                                                        "tags":  [
                                                                     "Touchpoint"
                                                                 ],
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/CreateTouchpointRequest"
                                                                                                                                }
                                                                                                                 },
                                                                                            "text/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/CreateTouchpointRequest"
                                                                                                                         }
                                                                                                          },
                                                                                            "application/*+json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/CreateTouchpointRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Success"
                                                                                  }
                                                                      }
                                                    },
                                           "get":  {
                                                       "tags":  [
                                                                    "Touchpoint"
                                                                ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Success"
                                                                                 }
                                                                     }
                                                   }
                                       },
                  "/api/touchpoints/{id}":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Touchpoint"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Success"
                                                                                      }
                                                                          }
                                                        },
                                                "put":  {
                                                            "tags":  [
                                                                         "Touchpoint"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/UpdateTouchpointRequest"
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/UpdateTouchpointRequest"
                                                                                                                             }
                                                                                                              },
                                                                                                "application/*+json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/UpdateTouchpointRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Success"
                                                                                      }
                                                                          }
                                                        }
                                            }
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateTouchpointRequest":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "name":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                          "type":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                          "configuration":  {
                                                                                                                "type":  "string",
                                                                                                                "nullable":  true
                                                                                                            },
                                                                                          "isActive":  {
                                                                                                           "type":  "boolean"
                                                                                                       }
                                                                                      },
                                                                       "additionalProperties":  false
                                                                   },
...
                                       "UpdateTouchpointRequest":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "name":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                          "configuration":  {
                                                                                                                "type":  "string",
                                                                                                                "nullable":  true
                                                                                                            },
                                                                                          "isActive":  {
                                                                                                           "type":  "boolean"
                                                                                                       }
                                                                                      },
                                                                       "additionalProperties":  false
                                                                   }
                                   }
                   }
}